fix tests, moved data accessors to value_source
