+++ corehq/motech/openmrs/openmrs_config.py
@@ -4,7 +4,6 @@ from __future__ import unicode_literals
 from corehq.motech.openmrs.finders import PatientFinder
 from corehq.motech.value_source import ValueSource
 from dimagi.ext.couchdbkit import (
-    DictProperty,
     DocumentSchema,
     ListProperty,
     SchemaDictProperty,
@@ -16,88 +15,6 @@ from dimagi.ext.couchdbkit import (
 class IdMatcher(DocumentSchema):
     case_property = StringProperty()
     identifier_type_id = StringProperty()
-
-
-class CaseProperty(ValueSource):
-    # Example "person_property" value::
-    #
-    #     {
-    #       "birthdate": {
-    #         "doc_type": "CaseProperty",
-    #         "case_property": "dob"
-    #       }
-    #     }
-    #
-    case_property = StringProperty()
-
-    def get_value(self, case_trigger_info):
-        return case_trigger_info.updates.get(self.case_property)
-
-
-class FormQuestion(ValueSource):
-    form_question = StringProperty()  # e.g. "/data/foo/bar"
-
-    def get_value(self, case_trigger_info):
-        return case_trigger_info.form_question_values.get(self.form_question)
-
-
-class ConstantString(ValueSource):
-    # Example "person_property" value::
-    #
-    #     {
-    #       "birthdate": {
-    #         "doc_type": "ConstantString",
-    #         "value": "Sep 7, 3761 BC"
-    #       }
-    #     }
-    #
-    value = StringProperty()
-
-    def get_value(self, case_trigger_info):
-        return self.value
-
-
-class CasePropertyMap(CaseProperty):
-    """
-    Maps case property values to OpenMRS values or concept UUIDs
-    """
-    # Example "person_attribute" value::
-    #
-    #     {
-    #       "00000000-771d-0000-0000-000000000000": {
-    #         "doc_type": "CasePropertyMap",
-    #         "case_property": "pill"
-    #         "value_map": {
-    #           "red": "00ff0000-771d-0000-0000-000000000000",
-    #           "blue": "000000ff-771d-0000-0000-000000000000",
-    #         }
-    #       }
-    #     }
-    #
-    value_map = DictProperty()
-
-    def get_value(self, case_trigger_info):
-        value = super(CasePropertyMap, self).get_value(case_trigger_info)
-        try:
-            return self.value_map[value]
-        except KeyError:
-            # We don't care if some CommCare answers are not mapped to OpenMRS concepts, e.g. when only the "yes"
-            # value of a yes-no question in CommCare is mapped to a concept in OpenMRS.
-            return None
-
-
-class FormQuestionMap(FormQuestion):
-    """
-    Maps form question values to OpenMRS values or concept UUIDs
-    """
-    value_map = DictProperty()
-
-    def get_value(self, case_trigger_info):
-        value = super(FormQuestionMap, self).get_value(case_trigger_info)
-        try:
-            return self.value_map[value]
-        except KeyError:
-            return None
 
 
 class OpenmrsCaseConfig(DocumentSchema):
